move ``_labels_equal()``, ValueError msg mistake
